CKANQuerry.jl

Documentation for CKANQuerry.jl

CKANQuerry.CKANConnection — Type

CKANConnection(url::String)
CKANConnection(url::String, apiversion::Int)

Creates a CKANConnection object with an API url, by first checking that the url is valid (i.e. return a status value of 200). An API version number (Int) can be specified. If not specified, the "connection" will use the latest version of the target API.

CKANQuerry.CKANDataset — Type

CKANDataset(connection::CKANConnection, id::String, metadata::Dict, resources::Union{Array{CKANResource}, Array{String}})

Creates a CKANDataset object. Datasets (i.e. packages) are the main objects in CKAN. CKANDataset objects usually contains one or more CKANResource objects.

CKANQuerry.CKANResource — Type

CKANResource(connection::CKANConnection, id::String, datasetid::String, metadata::Dict

Creates a CKANResource objetcs. Ressources are usually containes in CKAN datasets (i.e. packages).

Base.show — Method

Base.show(io::IO, connection::CKANConnection)

Print a CKANConnection to the given IO.

Base.show — Method

Base.show(io::IO, dataset::CKANDataset)

Print a CKANDataset to the given IO.

Base.show — Method

Base.show(io::IO, resource::CKANResource)

Print a CKANResource to the given IO.

CKANQuerry.generate_request_query — Method

generate_request_query(parameters::Pair...)

Copied from Mangal.jl, formats a query string for use in a url, from a set of Pairs.

CKANQuerry.http_get — Method

http_get(url::String)

Perform GET request on url and return the response.

CKANQuerry.make_api_url — Method

make_api_url(root::String)
make_api_url(root::String, apiversion::Int)

Produce a GETable api url. Root is the base url, apiversion is the version of the api to use. If apiversion is not specified, the default behavior when a querry is performed is to use the latest version.

CKANQuerry.parse_body — Method

parse_body(response::HTTP.Messages.Response)

Parse the body of a HTTP response.

CKANQuerry.querry_api — Method

querry_api(url::String, action::String, parameters::Dict)

Querry an API with the given parameters, for a given CKAN action.

CKANQuerry.shorten_url — Function

shorten_url(url::String, squeeze::Int=20)

Shorten the url for printing.

CKANQuerry.test_url — Method

test_url(url::String)

Test url validity.